agent/data_ingestion/pyproject.toml
[project]
name = "data-ingestion-pipeline"
version = "0.1.0"
description = "Data ingestion pipeline for Mozilla data-docs into Vertex AI Search"
readme = "README.md"
requires-python = ">=3.9, <=3.13"
dependencies = [
    "backoff>=2.2.0",
    "google-cloud-aiplatform>=1.80.0",
    "google-cloud-pipeline-components>=2.19.0",
    "google-cloud-storage>=2.14.0",
    "kfp>=1.4.0",
    "markdown>=3.5.0",
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["data_ingestion_pipeline"]